+++ electron/src/api/github/login.ts
@@ -1,4 +1,4 @@
-import fetch from 'node-fetch';
+import fetch from 'cross-fetch';
 
 type Credentials = {
   clientId: string;
