+++ package.json
@@ -33,7 +33,6 @@
     "classnames": "^2.3.1",
     "destyle.css": "^2.0.2",
     "jotai": "^1.1.3",
-    "node-fetch": "^2.6.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-relay": "^11.0.2",
@@ -58,6 +57,7 @@
     "@vanilla-extract/vite-plugin": "^2.0.0",
     "@vitejs/plugin-react-refresh": "^1.3.1",
     "babel-plugin-relay": "^11.0.2",
+    "cross-fetch": "^3.1.4",
     "dotenv": "^10.0.0",
     "electron": "^13.1.5",
     "electron-builder": "^22.11.7",
