--- renderer/src/graphql/mapper/index.ts
@@ -1,1 +1,0 @@
-export * from './pullRequest';
